# src/main/java/org/example/cuidadodemascotas/servicemicroservice/apis/controller/InvoiceApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.cuidadodemascotas.servicemicroservice.apis.controller;

import org.example.cuidadodemascotas.servicemicroservice.apis.dto.InvoiceRequestDTO;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.InvoiceResponseDTO;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.InvoiceResult;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-20T03:45:32.928054400-03:00[America/Asuncion]")
@Validated
@Tag(name = "Invoice", description = "the Invoice API")
public interface InvoiceApi {

    /**
     * POST /invoices : Crear una nueva factura
     *
     * @param invoiceRequestDTO  (required)
     * @return Factura creada exitosamente (status code 201)
     *         or Solicitud inválida (status code 400)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "createInvoice",
        summary = "Crear una nueva factura",
        tags = { "Invoice" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Factura creada exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = InvoiceResponseDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida"),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/invoices",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<InvoiceResponseDTO> createInvoice(
        @Parameter(name = "InvoiceRequestDTO", description = "", required = true) @Valid @RequestBody InvoiceRequestDTO invoiceRequestDTO
    );


    /**
     * DELETE /invoices/{id} : Eliminar lógicamente una factura
     *
     * @param id  (required)
     * @return Factura eliminada exitosamente (status code 204)
     *         or Factura no encontrada (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "deleteInvoice",
        summary = "Eliminar lógicamente una factura",
        tags = { "Invoice" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Factura eliminada exitosamente"),
            @ApiResponse(responseCode = "404", description = "Factura no encontrada"),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/invoices/{id}"
    )
    
    ResponseEntity<Void> deleteInvoice(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /invoices/{id} : Obtener una factura por ID
     *
     * @param id  (required)
     * @return Factura obtenida exitosamente (status code 200)
     *         or Factura no encontrada (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getInvoiceById",
        summary = "Obtener una factura por ID",
        tags = { "Invoice" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Factura obtenida exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = InvoiceResponseDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Factura no encontrada"),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/invoices/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<InvoiceResponseDTO> getInvoiceById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /invoices : Obtener todas las facturas con paginación y filtros
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @param reservationId Filtrar por ID de reserva (optional)
     * @return Facturas obtenidas exitosamente (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getInvoices",
        summary = "Obtener todas las facturas con paginación y filtros",
        tags = { "Invoice" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Facturas obtenidas exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = InvoiceResult.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/invoices",
        produces = { "application/json" }
    )
    
    ResponseEntity<InvoiceResult> getInvoices(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size,
        @Parameter(name = "reservationId", description = "Filtrar por ID de reserva", in = ParameterIn.QUERY) @Valid @RequestParam(value = "reservationId", required = false) Long reservationId
    );


    /**
     * PUT /invoices/{id} : Actualizar una factura existente
     *
     * @param id  (required)
     * @param invoiceRequestDTO  (required)
     * @return Factura actualizada exitosamente (status code 200)
     *         or Factura no encontrada (status code 404)
     *         or Solicitud inválida (status code 400)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "updateInvoice",
        summary = "Actualizar una factura existente",
        tags = { "Invoice" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Factura actualizada exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = InvoiceResponseDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Factura no encontrada"),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida"),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/invoices/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<InvoiceResponseDTO> updateInvoice(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "InvoiceRequestDTO", description = "", required = true) @Valid @RequestBody InvoiceRequestDTO invoiceRequestDTO
    );

}