# src/main/java/org/example/cuidadodemascotas/servicemicroservice/apis/controller/PaymentDetailApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.cuidadodemascotas.servicemicroservice.apis.controller;

import org.example.cuidadodemascotas.servicemicroservice.apis.dto.PaymentDetailRequestDTO;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.PaymentDetailResponseDTO;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.PaymentDetailResult;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-20T03:45:32.928054400-03:00[America/Asuncion]")
@Validated
@Tag(name = "PaymentDetail", description = "the PaymentDetail API")
public interface PaymentDetailApi {

    /**
     * POST /payment-details : Crear un nuevo detalle de pago
     *
     * @param paymentDetailRequestDTO  (required)
     * @return Detalle de pago creado exitosamente (status code 201)
     *         or Solicitud inválida (status code 400)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "createPaymentDetail",
        summary = "Crear un nuevo detalle de pago",
        tags = { "PaymentDetail" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Detalle de pago creado exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PaymentDetailResponseDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida"),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/payment-details",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<PaymentDetailResponseDTO> createPaymentDetail(
        @Parameter(name = "PaymentDetailRequestDTO", description = "", required = true) @Valid @RequestBody PaymentDetailRequestDTO paymentDetailRequestDTO
    );


    /**
     * GET /payment-details : Obtener todos los detalles de pago
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Detalles de pago obtenidos exitosamente (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getPaymentDetails",
        summary = "Obtener todos los detalles de pago",
        tags = { "PaymentDetail" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Detalles de pago obtenidos exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PaymentDetailResult.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/payment-details",
        produces = { "application/json" }
    )
    
    ResponseEntity<PaymentDetailResult> getPaymentDetails(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    );

}