# src/main/resources/config-openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice and Payments Microservice API
  description: Microservicio para gestión de facturas y pagos
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Servicio de Facturas y Pagos
paths:
...
  /invoices:
    get:
      tags:
        - Invoice
      summary: Obtener todas las facturas con paginación y filtros
      operationId: getInvoices
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: reservationId
          in: query
          required: false
          description: Filtrar por ID de reserva
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Facturas obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResult'
        '500':
          description: Error interno del servidor

    post:
      tags:
        - Invoice
      summary: Crear una nueva factura
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRequestDTO'
      responses:
        '201':
          description: Factura creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponseDTO'
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor
  /invoices/{id}:
    get:
      tags:
        - Invoice
      summary: Obtener una factura por ID
      operationId: getInvoiceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Factura obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponseDTO'
        '404':
          description: Factura no encontrada
        '500':
          description: Error interno del servidor

    put:
      tags:
        - Invoice
      summary: Actualizar una factura existente
      operationId: updateInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRequestDTO'
      responses:
        '200':
          description: Factura actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponseDTO'
        '404':
          description: Factura no encontrada
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor

    delete:
      tags:
        - Invoice
      summary: Eliminar lógicamente una factura
      operationId: deleteInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Factura eliminada exitosamente
        '404':
          description: Factura no encontrada
        '500':
          description: Error interno del servidor

  # ==================== PAYMENT METHODS ====================
...
  /payment-details:
    get:
      tags:
        - PaymentDetail
      summary: Obtener todos los detalles de pago
      operationId: getPaymentDetails
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Detalles de pago obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDetailResult'
        '500':
          description: Error interno del servidor

    post:
      tags:
        - PaymentDetail
      summary: Crear un nuevo detalle de pago
      operationId: createPaymentDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentDetailRequestDTO'
      responses:
        '201':
          description: Detalle de pago creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDetailResponseDTO'
        '400':
          description: Solicitud inválida
        '500':
          description: Error interno del servidor
components:
  schemas:
...
    InvoiceRequestDTO:
      type: object
      required:
        - reservationId
        - totalAmount
      properties:
        reservationId:
          type: integer
          format: int64
        issueDate:
          type: string
          format: date-time
        totalAmount:
          type: number
          format: double
    InvoiceResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        active:
          type: boolean
        reservationId:
          type: integer
          format: int64
        issueDate:
          type: string
          format: date-time
        totalAmount:
          type: number
          format: double
    InvoiceResult:
      type: object
      properties:
        totalCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceResponseDTO'

    # ==================== PAYMENT METHOD ====================
...
    PaymentDetailRequestDTO:
      type: object
      required:
        - invoiceId
        - paymentMethodId
        - appliedAmount
      properties:
        invoiceId:
          type: integer
          format: int64
        paymentMethodId:
          type: integer
          format: int64
        appliedAmount:
          type: number
          format: double
    PaymentDetailResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        active:
          type: boolean
        invoiceId:
          type: integer
          format: int64
        paymentMethodId:
          type: integer
          format: int64
        appliedAmount:
          type: number
          format: double
    PaymentDetailResult:
      type: object
      properties:
        totalCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentDetailResponseDTO'